bot/requirements.txt
annotated-types==0.7.0
cffi==1.17.1
click==8.1.8
ffmpeg-python==0.2.0
llvmlite==0.44.0
numba==0.61.0
numpy==2.1.3
openai==1.65.1
openai-whisper==20240930
asyncpg==0.28.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-telegram-bot==21.10
regex==2024.11.6
sentry-sdk==2.29.1
soundfile==0.13.1
sqlalchemy==2.0.43
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
greenlet==3.2.4
passlib[bcrypt]==1.7.4
redis==6.4.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
alembic==1.16.4
requests==2.32.3
psycopg==3.2.10
psycopg-binary==3.2.10